Aggiunto db context EF per lista corsi

+++ Models/Entities/Lessons.cs
@@ -6,7 +6,7 @@ namespace MyCourse.Models.Entities
     public partial class Lesson
     {
         public long Id { get; set; }
-        public long CourseId { get; set; }
+        public int CourseId { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
         public string Duration { get; set; }
